csv logger moved to utils
