+++ src/rluni/configs/robot/unicycle.yaml
@@ -9,7 +9,7 @@ RobotSystem:
   imu_config2: "configs/imu/full-robot-imu2.yaml"
   rlmodel_path: "configs/rlmodels/broomy-balance-export.onnx"  # Relative path to RL model file
 
-  csv_logging: False           # Enable CSV logging
+  csv_logging: True        # Enable CSV logging
 
   pitch_wheel_radius: 0.05395
